rework PID

+++ Core/Inc/control/coil_controller.h
@@ -32,8 +32,8 @@ typedef struct
     PWMController_t PWM[NUMBER_OF_THERMISTORS];
     OperationMode_t mode;
     ControlReference_t control_reference;
-    int16_t u;
-    uint16_t u_saturated;
+    float u;
+    float u_saturated;
     uint16_t u_max;
     uint16_t u_min;
 } CoilController_t;
