+++ Core/Inc/control/pid.h
@@ -27,7 +27,7 @@ typedef struct
     IIRfilter_t error_difference;
 } PID_t;
 
-int16_t PID_update(PID_t *this, float error, int16_t control_difference);
+float PID_update(PID_t *this, float error, int16_t control_difference);
 void PID_update_with_error_injection(PID_t *this, float error);
 void PID_reset(PID_t *this);
